local read receipts for notifications
