+++ src/app/MenuMobile.svelte
@@ -9,7 +9,8 @@
   import MenuMobileItem from "src/app/MenuMobileItem.svelte"
   import {slowConnections, menuIsOpen} from "src/app/state"
   import {router} from "src/app/util/router"
-  import {env, hasNewMessages, hasNewNotifications} from "src/engine"
+  import {env, hasNewMessages} from "src/engine"
+  import {hasNewNotifications} from "src/domain/notifications"
 
   const closeSubMenu = () => {
     subMenu = null
