+++ src/app/views/NotificationSectionReactions.svelte
@@ -1,13 +1,17 @@
 <script lang="ts">
   import {onMount} from "svelte"
   import {derived} from "svelte/store"
-  import {partition, max, pushToMapKey} from "@welshman/lib"
+  import {max, pushToMapKey} from "@welshman/lib"
   import type {TrustedEvent} from "@welshman/util"
-  import {getAncestorTagValues, SEEN_GENERAL, REACTION} from "@welshman/util"
+  import {getAncestorTagValues} from "@welshman/util"
   import {formatTimestampAsDate} from "src/util/misc"
   import NotificationList from "src/app/views/NotificationList.svelte"
   import NotificationReactions from "src/app/views/NotificationReactions.svelte"
-  import {reactionNotifications, unreadReactionNotifications, markAsSeen} from "src/engine"
+  import {
+    reactionNotifications,
+    setChecked,
+    unreadReactionNotifications,
+  } from "src/domain/notifications"
 
   export let limit
 
@@ -32,26 +36,19 @@
     })
   })
 
-  let loading = false
-
   onMount(() => {
     const tracked = new Set()
+    setChecked(["reactions", "zaps"])
 
     const unsub = unreadReactionNotifications.subscribe(async events => {
       const untracked = events.filter(e => !tracked.has(e.id))
 
-      if (!loading && untracked.length > 0) {
+      if (untracked.length > 0) {
         for (const id of untracked) {
           tracked.add(id)
         }
 
-        const [reactions, zaps] = partition(e => e.kind === REACTION, $reactionNotifications)
-
-        loading = true
-
-        await markAsSeen(SEEN_GENERAL, {reactions, zaps})
-
-        loading = false
+        setChecked(["reactions", "zaps"])
       }
     })
