+++ src/app/views/Notifications.svelte
@@ -6,12 +6,8 @@
   import NotificationSectionMain from "src/app/views/NotificationSectionMain.svelte"
   import NotificationSectionReactions from "src/app/views/NotificationSectionReactions.svelte"
   import {router} from "src/app/util/router"
-  import {
-    sessionWithMeta,
-    loadNotifications,
-    unreadMainNotifications,
-    unreadReactionNotifications,
-  } from "src/engine"
+  import {sessionWithMeta, loadNotifications} from "src/engine"
+  import {unreadMainNotifications, unreadReactionNotifications} from "src/domain/notifications"
 
   const allTabs = ["Mentions & Replies", "Reactions"]
